+++ fj-core-jvfs/src/main/java/org/fugerit/java/core/jvfs/db/daogen/def/facade/DbJvfsFileFinder.java
@@ -1,6 +1,6 @@
 package org.fugerit.java.core.jvfs.db.daogen.def.facade;
 
-import org.fugerit.java.core.db.daogen.BaseIdFinder;
+import org.fugerit.java.core.db.daogen.IdFinderNG;
 import org.fugerit.java.core.jvfs.db.daogen.model.ModelDbJvfsFile;
 
 // custom import start ( code above here will be overwritten )
@@ -15,29 +15,10 @@ import org.fugerit.java.core.jvfs.db.daogen.model.ModelDbJvfsFile;
  * // custom code start ( code above here will be overwritten )
  * // custom code end ( code below here will be overwritten )
  */
-public class DbJvfsFileFinder extends BaseIdFinder {
+public class DbJvfsFileFinder extends IdFinderNG {
 
 	// custom code start ( code above here will be overwritten )
-	
-	// code added to setup a basic conditional serialization - START
-	
-	private void writeObject(java.io.ObjectOutputStream out) throws java.io.IOException {
-		// this class is conditionally serializable, depending on contained object
-		// you are encouraged to handle special situation using this method
-		out.defaultWriteObject();
-	}
-
-	private void readObject(java.io.ObjectInputStream in) throws java.io.IOException, ClassNotFoundException {
-		// this class is conditionally serializable, depending on contained object
-		// you are encouraged to handle special situation using this method
-		in.defaultReadObject();
-	}
-	
-	// code added to setup a basic conditional serialization - END
-	
 	// custom code end ( code below here will be overwritten )
-
-	private static final long serialVersionUID = 4317877389001250786L;
 
 	private ModelDbJvfsFile model;